enormous refactroing finished
