+++ functions/user_manager.js
