+++ server_storage/guild_list.json
@@ -1,1 +1,1 @@
-{"710746690650374208":{"portal_list":{"717403805712646215":{"creator_id":"228666893068795905","regex_portal":"haek","regex_voice":"G$#-P$member_count | $status_list","no_bots":false,"limit_portal":0,"time_to_live":0,"refresh_rate":0,"locale":"gr","voice_list":{},"force_update":false,"last_update":1591112795710},"717878219495178242":{"creator_id":"228666893068795905","regex_portal":"neame","regex_voice":"G$#-P$member_count | $status_list","no_bots":false,"limit_portal":0,"time_to_live":0,"refresh_rate":0,"locale":"gr","voice_list":{},"force_update":false,"last_update":1591225905030}},"url_list":{},"role_list":{}},"228667314252283904":{"portal_list":{},"url_list":{},"role_list":{}},"705112209985896529":{"portal_list":{"717856041219653679":{"creator_id":"228666893068795905","regex_portal":"newone","regex_voice":"G$#-P$member_count | $status_list","no_bots":false,"limit_portal":0,"time_to_live":0,"refresh_rate":0,"locale":"gr","voice_list":{},"force_update":false,"last_update":1591220617307},"717856223273418863":{"creator_id":"228666893068795905","regex_portal":"ena","regex_voice":"G$#-P$member_count | $status_list","no_bots":false,"limit_portal":0,"time_to_live":0,"refresh_rate":0,"locale":"gr","voice_list":{},"force_update":false,"last_update":1591220660661},"717856372175274016":{"creator_id":"228666893068795905","regex_portal":"enas","regex_voice":"G$#-P$member_count | $status_list","no_bots":false,"limit_portal":0,"time_to_live":0,"refresh_rate":0,"locale":"gr","voice_list":{},"force_update":false,"last_update":1591220696216}},"url_list":{},"role_list":{}}}
+{"228667314252283904":{"portal_list":{},"url_list":{},"role_list":{}},"705112209985896529":{"portal_list":{"717856041219653679":{"creator_id":"228666893068795905","regex_portal":"newone","regex_voice":"G$#-P$member_count | $status_list","no_bots":false,"limit_portal":0,"time_to_live":0,"refresh_rate":0,"locale":"gr","voice_list":{},"force_update":false,"last_update":1591220617307},"717856223273418863":{"creator_id":"228666893068795905","regex_portal":"ena","regex_voice":"G$#-P$member_count | $status_list","no_bots":false,"limit_portal":0,"time_to_live":0,"refresh_rate":0,"locale":"gr","voice_list":{},"force_update":false,"last_update":1591220660661},"717856372175274016":{"creator_id":"228666893068795905","regex_portal":"enas","regex_voice":"G$#-P$member_count | $status_list","no_bots":false,"limit_portal":0,"time_to_live":0,"refresh_rate":0,"locale":"gr","voice_list":{},"force_update":false,"last_update":1591220696216}},"url_list":{},"role_list":{}},"710746690650374208":{"portal_list":{"717897576975826954":{"creator_id":"228666893068795905","regex_portal":"ehle","regex_voice":"G$#-P$member_count | $status_list","no_bots":false,"limit_portal":0,"time_to_live":0,"refresh_rate":0,"locale":"gr","voice_list":{"717897582965424200":{"creator_id":"228666893068795905","regex":"G$#-P$member_count | $status_list","no_bots":false,"time_to_live":0,"refresh_rate":0,"locale":"gr","force_update":1,"last_update":1591230521501}},"force_update":false,"last_update":1591230520072}},"url_list":{},"role_list":{}}}
